perf(voice): single-pass TTS text cleanup in for_voice
for_voice runs on every spoken assistant reply and was a multi-pass, allocation-heavy pipeline. clean_for_tts chained twelve whole-buffer String::replace calls (one allocation each); strip_raw_urls built a Vec<&str> then join'd it; truncate_for_voice collected a Vec<String> of per-sentence clones then join'd. This collapses the order-independent rewrites into single passes while preserving the order-sensitive ones exactly:

- clean_for_tts: keep the four ordered multi-char replaces (..., ' - ', ' — ', ' – ' — their overlapping-space semantics are order-dependent) and the trailing possessive 's->s, but fold the seven single-char rules (parens -> ', ', drop []{}") into one char-scan pass (12 allocations -> 6).
- strip_raw_urls: stream kept tokens straight into the result String instead of Vec + join.
- truncate_for_voice: accumulate sentences into one String with a running count instead of Vec<String> + join, dropping the per-sentence clones.

Behavior is byte-identical: verified by a differential fuzz harness (400k random inputs plus a markdown/punctuation corpus) comparing the new output against the original, and by the existing 16 tests. Those tests only exercise the public for_voice API, so they move to tests/voice_format_test.rs (integration); tests/voice_format_bench.rs adds an ignored microbench. Measured ~1.2x faster on an x86 dev box; Jetson numbers in the PR body. Contributes under #402 (performance bucket).
