+++ crates/genie-core/src/voice/format.rs
@@ -132,20 +132,26 @@ fn strip_links(text: &str) -> String {
 }
 
 fn strip_raw_urls(text: &str) -> String {
-    text.split_whitespace()
-        .filter(|token| {
-            let trimmed = token.trim_matches(|c: char| {
-                matches!(
-                    c,
-                    '(' | ')' | '[' | ']' | '{' | '}' | ',' | ';' | ':' | '"' | '\''
-                )
-            });
-            !(trimmed.starts_with("http://")
-                || trimmed.starts_with("https://")
-                || trimmed.starts_with("www."))
-        })
-        .collect::<Vec<_>>()
-        .join(" ")
+    let mut result = String::with_capacity(text.len());
+    for token in text.split_whitespace() {
+        let trimmed = token.trim_matches(|c: char| {
+            matches!(
+                c,
+                '(' | ')' | '[' | ']' | '{' | '}' | ',' | ';' | ':' | '"' | '\''
+            )
+        });
+        if trimmed.starts_with("http://")
+            || trimmed.starts_with("https://")
+            || trimmed.starts_with("www.")
+        {
+            continue;
+        }
+        if !result.is_empty() {
+            result.push(' ');
+        }
+        result.push_str(token);
+    }
+    result
 }
 
 /// Normalize whitespace: collapse multiple spaces, trim.
@@ -200,7 +206,8 @@ fn truncate_for_voice(text: &str, max_sentences: usize) -> String {
     }
 
     let chars: Vec<char> = text.chars().collect();
-    let mut sentences = Vec::new();
+    let mut result = String::with_capacity(text.len());
+    let mut sentence_count = 0;
     let mut current = String::new();
 
     for (i, &ch) in chars.iter().enumerate() {
@@ -227,21 +234,28 @@ fn truncate_for_voice(text: &str, max_sentences: usize) -> String {
         };
 
         if ends_sentence {
-            sentences.push(current.trim().to_string());
+            if !result.is_empty() {
+                result.push(' ');
+            }
+            result.push_str(current.trim());
+            sentence_count += 1;
             current.clear();
-            if sentences.len() >= max_sentences {
+            if sentence_count >= max_sentences {
                 break;
             }
         }
     }
 
     // Include the trailing fragment if we still have room.
-    let trailing = current.trim().to_string();
-    if !trailing.is_empty() && sentences.len() < max_sentences {
-        sentences.push(trailing);
+    let trailing = current.trim();
+    if !trailing.is_empty() && sentence_count < max_sentences {
+        if !result.is_empty() {
+            result.push(' ');
+        }
+        result.push_str(trailing);
     }
 
-    sentences.join(" ")
+    result
 }
 
 fn is_ascii_terminator(ch: char) -> bool {
@@ -269,166 +283,20 @@ fn ends_with_abbreviation(current: &str) -> bool {
 
 /// Clean special characters that TTS engines handle poorly.
 fn clean_for_tts(text: &str) -> String {
-    text.replace("...", ", ")
+    let pre = text
+        .replace("...", ", ")
         .replace(" - ", ", ")
         .replace(" — ", ", ")
-        .replace(" – ", ", ")
-        .replace("(", ", ")
-        .replace(")", ", ")
-        .replace("[", "")
-        .replace("]", "")
-        .replace("{", "")
-        .replace("}", "")
-        .replace("\"", "")
-        .replace("'s", "s") // possessive sounds weird with some TTS
-}
+        .replace(" – ", ", ");
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn strip_bold_and_italic() {
-        assert_eq!(
-            for_voice("This is **bold** and *italic*"),
-            "This is bold and italic"
-        );
+    let mut result = String::with_capacity(pre.len());
+    for ch in pre.chars() {
+        match ch {
+            '(' | ')' => result.push_str(", "),
+            '[' | ']' | '{' | '}' | '"' => {}
+            _ => result.push(ch),
+        }
     }
 
-    #[test]
-    fn strip_markdown_headers() {
-        assert_eq!(for_voice("## Weather\nIt's sunny."), "Weather Its sunny.");
-    }
-
-    #[test]
-    fn strip_bullet_points() {
-        let input = "Here's what I found:\n- Item one\n- Item two\n- Item three";
-        let output = for_voice(input);
-        assert!(output.contains("Item one"));
-        assert!(!output.contains("- "));
-    }
-
-    #[test]
-    fn strip_code_blocks() {
-        let input = "Here's the code:\n```\nlet x = 5;\n```\nThat's it.";
-        let output = for_voice(input);
-        assert!(!output.contains("let x"));
-        assert!(output.contains("Thats it"));
-    }
-
-    #[test]
-    fn strip_links() {
-        let input = "Check [this guide](https://example.com) for details.";
-        let output = for_voice(input);
-        assert!(output.contains("this guide"));
-        assert!(!output.contains("https://"));
-    }
-
-    #[test]
-    fn strip_raw_urls_from_plain_text() {
-        let input = "Top result: ESP32-C6 supports Thread. https://example.com/thread";
-        let output = for_voice(input);
-        assert!(output.contains("ESP32-C6 supports Thread"));
-        assert!(!output.contains("https://"));
-    }
-
-    #[test]
-    fn truncate_long_response() {
-        let input =
-            "First sentence. Second sentence. Third sentence. Fourth sentence. Fifth sentence.";
-        let output = for_voice(input);
-        assert!(output.contains("First"));
-        assert!(output.contains("Third"));
-        assert!(!output.contains("Fourth"));
-    }
-
-    #[test]
-    fn truncate_handles_chinese_punctuation() {
-        let input = "第一句。第二句！第三句？第四句。";
-        let output = for_voice(input);
-        assert!(output.contains("第一句"));
-        assert!(output.contains("第三句"));
-        assert!(!output.contains("第四句"));
-    }
-
-    #[test]
-    fn clean_special_chars() {
-        let input = "The temperature is 72°F (about 22°C)...nice!";
-        let output = for_voice(input);
-        assert!(!output.contains("("));
-        assert!(!output.contains("..."));
-    }
-
-    #[test]
-    fn empty_input() {
-        assert_eq!(for_voice(""), "");
-    }
-
-    #[test]
-    fn already_clean() {
-        assert_eq!(for_voice("The lights are on."), "The lights are on.");
-    }
-
-    #[test]
-    fn decimal_numbers_are_not_split() {
-        let output = for_voice("It is 72.5 degrees outside and feels warm.");
-        assert!(
-            output.contains("72.5"),
-            "decimal must stay intact, got {output:?}"
-        );
-        assert!(
-            !output.contains("72. 5"),
-            "decimal must not be split, got {output:?}"
-        );
-    }
-
-    #[test]
-    fn version_strings_are_not_split() {
-        let output = for_voice("Version v1.0.0 is installed and ready.");
-        assert!(
-            output.contains("v1.0.0"),
-            "version dots must stay intact, got {output:?}"
-        );
-    }
-
-    #[test]
-    fn abbreviations_do_not_end_a_sentence() {
-        let output = for_voice("Dinner is at 6 p.m. today in the kitchen.");
-        // "p.m." must survive intact — no split at "p." or "p.m. " mid-clause.
-        assert!(
-            output.contains("p.m."),
-            "abbreviation must stay intact, got {output:?}"
-        );
-        assert!(
-            !output.contains("p. m."),
-            "abbreviation must not be split, got {output:?}"
-        );
-        assert!(
-            output.contains("today in the kitchen"),
-            "clause after the abbreviation must remain, got {output:?}"
-        );
-    }
-
-    #[test]
-    fn cap_counts_real_sentences_not_decimal_fragments() {
-        // Before the fix, "3.50" split into two fake sentences and burned the
-        // 3-sentence budget, dropping "All systems normal." entirely.
-        let output =
-            for_voice("The CPU is at 3.50 GHz now. Memory usage is fine. All systems normal.");
-        assert!(output.contains("3.50 GHz"), "got {output:?}");
-        assert!(output.contains("Memory usage is fine"), "got {output:?}");
-        assert!(
-            output.contains("All systems normal"),
-            "the third real sentence must not be dropped, got {output:?}"
-        );
-    }
-
-    #[test]
-    fn chinese_punctuation_still_segments_immediately() {
-        // CJK terminators have no trailing space; they must still split.
-        let output = for_voice("第一句。第二句！第三句？第四句。");
-        assert!(output.contains("第一句"));
-        assert!(output.contains("第三句"));
-        assert!(!output.contains("第四句"));
-    }
+    result.replace("'s", "s")
 }
